+++ docs/save.txt
@@ -1,2 +1,1 @@
-T>1>todo1
-D>1>dl>2020-02-02T16:23
+T>0>task
