+++ src/main/java/Duke/ByeCommand.java
@@ -1,3 +1,5 @@
+package Duke;
+
 public class ByeCommand implements ICommand {
 
     public void execute(TaskManager tm, Ui ui, Storage storage) {
