+++ src/main/java/Duke/DoneCommand.java
@@ -1,22 +1,25 @@
-public class DeleteCommand implements ICommand {
+package Duke;
+
+public class DoneCommand implements ICommand {
 
     String taskIndex;
 
-    public DeleteCommand(String taskIndex) {
+    public DoneCommand(String taskIndex) {
         this.taskIndex = taskIndex;
     }
 
     public void execute(TaskManager tm, Ui ui, Storage storage) {
         try {
-            Task deletedTask = tm.deleteTask(taskIndex);
-            if (deletedTask == null) {
+            Task completedTask = tm.completeTask(taskIndex);
+            if (completedTask == null) {
                 ui.printInvalidIndexMessage();
             } else {
-                ui.printTaskDeletion(deletedTask, tm.getTasks().size());
+                ui.printTaskCompletion(completedTask, tm.getTasks().size());
                 storage.updateSave(tm.getTasks());
             }
         } catch (DukeException.InvalidInputException e) {
             ui.printErrorMessage(e);
         }
     }
+
 }
