+++ src/main/java/Duke/ICommand.java
@@ -1,3 +1,5 @@
+package Duke;
+
 public interface ICommand {
 
     public abstract void execute(TaskManager tm, Ui ui, Storage storage);
