+++ src/main/java/Duke/Parser.java
@@ -1,3 +1,5 @@
+package Duke;
+
 import java.util.Scanner;
 
 public class Parser {
