+++ src/main/java/Duke/UnknownCommand.java
@@ -1,3 +1,5 @@
+package Duke;
+
 public class UnknownCommand implements ICommand {
     @Override
     public void execute(TaskManager tm, Ui ui, Storage storage) {
