+++ text-ui-test/EXPECTED.TXT
@@ -7,74 +7,74 @@ Hello I'm
 What can I do for you?
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-Duke says: Here is your list of tasks :)
+Duke.Duke says: Here is your list of tasks :)
 Looks like you don't have any pending tasks! Must be nice (-_-;)
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-Duke says: You can't add a task with no name
+Duke.Duke says: You can't add a task with no name
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-Duke says: Please include a time
+Duke.Duke says: Please include a time
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-Duke says: Please include a deadline
+Duke.Duke says: Please include a deadline
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-Duke says: I've added the task:
+Duke.Duke says: I've added the task:
      #ToDo (not done) get some sleep
 You now have 1 tasks, jiayouz!
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-Duke says: Please include a time
+Duke.Duke says: Please include a time
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-Duke says: I've added the task:
+Duke.Duke says: I've added the task:
      #Event (not done) dinner (thurs)
 You now have 2 tasks, jiayouz!
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-Duke says: Please include a deadline
+Duke.Duke says: Please include a deadline
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-Duke says: I've added the task:
+Duke.Duke says: I've added the task:
      #Deadline (not done) homework (monday)
 You now have 3 tasks, jiayouz!
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-Duke says: Here is your list of tasks :)
+Duke.Duke says: Here is your list of tasks :)
 1.#ToDo (not done) get some sleep
 2.#Event (not done) dinner (thurs)
 3.#Deadline (not done) homework (monday)
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-Duke says: You have completed the task get some sleep. Well done!
+Duke.Duke says: You have completed the task get some sleep. Well done!
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-Duke says: You have completed the task dinner. Well done!
+Duke.Duke says: You have completed the task dinner. Well done!
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-Duke says: You have completed the task homework. Well done!
+Duke.Duke says: You have completed the task homework. Well done!
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-Duke says: Here is your list of tasks :)
+Duke.Duke says: Here is your list of tasks :)
 1.#ToDo (done) get some sleep
 2.#Event (done) dinner (thurs)
 3.#Deadline (done) homework (monday)
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-Duke says: Sorry I don't understand what that means
+Duke.Duke says: Sorry I don't understand what that means
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-Duke says: Here is your list of tasks :)
+Duke.Duke says: Here is your list of tasks :)
 1.#ToDo (done) get some sleep
 2.#Event (done) dinner (thurs)
 3.#Deadline (done) homework (monday)
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-Duke says: You have deleted the task get some sleep
+Duke.Duke says: You have deleted the task get some sleep
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-Duke says: Here is your list of tasks :)
+Duke.Duke says: Here is your list of tasks :)
 1.#Event (done) dinner (thurs)
 2.#Deadline (done) homework (monday)
 ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
